Skip flow lines to and from the same point
Modifies the logic to ensure that there are no flow lines in the output which start and end at the same point. This is done by adding a simple check whether the previous and current cell and IDs are the same.

This commit also moves the instantiation of the weight and prev_cell_id outside of the loop, as there is no reason to repeatedly instantiate these values.